TFG-prediction refactor location of the method
