+++ code/repositories/raziel_repository.py
@@ -1,3 +1,4 @@
+from models.exceptions import NoCorrectColumnsException
 from models.small_models import Sex
 from repositories.abstract_repository import AbstractRepository, ListParams
 from models.raziel_model import Raziel
@@ -38,3 +39,18 @@ class RazielRepository(AbstractRepository):
     def get_one(self, id: int) -> object:
         data = super().get_one('id', id)
         return self._transform_to_model(data) if data else None
+
+    def prepare_and_gruping_dataframe(self, params: ListParams, var1: str, var2: str):
+        df = self.filter_dataframe(params)
+        self._check_params(df, var1, var2)
+        cols = df.columns.tolist()
+        for i in [6, 7, 8, 9, 10, 11, 12, 13, 14]:
+            df = df.drop(labels=cols[i], axis=1)
+        df = df.groupby([var1])[var2].sum()
+        return df
+
+    @staticmethod
+    def _check_params(dataframe, var1: str, var2: str):
+        cols = dataframe.columns.tolist()
+        if var1 not in cols or var2 not in cols:
+            raise NoCorrectColumnsException('This columns does not exist in dataframe')
